removed old unusued stuff
